Handle when current file is CSHTML in C# project
Restructure guess language algorithm. Previously it tried to make its
first guess on the basis of what filetype the open file is, only falling
back to looking for project files if the current filetype was blank. The
purpose of this was performance. However, this no longer makes sense,
now that I am moving toward having project-file specific algorithms. In
other words, even if the current file is a C# file, the current C#
algorithm will not work if it can't find a solution file. That being the
case, I have privileged checking for project files in the guessing
algorithm.